requirements.txt
flask
flask-socketio
opencv-python
tensorflow
numpy
mediapipe
python-dotenv
SQLAlchemy
